Run tests-service as autopkgtest
Drop the obsolete and rather unnecessary setup from start-service (this is a
system integration test, stuff shoudl just run after installation or get dbus
activated). In lieu of an actual integration test that uses the installed
package, run test-service. This won't check the installed indicator-session
package, but at least verify that it still builds and runs, thus find
regressions in the dependencies.